+++ lib/pocketbase_helpers.dart
@@ -1,12 +1,7 @@
-/// Support for doing something awesome.
-///
-/// More dartdocs go here.
 library;
 
-export 'src/models.dart';
+export 'src/typed_result_list.dart';
 export 'src/collection_helper.dart';
 export 'src/base_helper.dart';
 export 'src/helper_utils.dart';
 export 'src/pocketbase_record.dart';
-
-// TODO: Export any libraries intended for clients of this package.